LayoutTests/platform/chromium-mac-snowleopard/media/video-poster-blocked-by-willsendrequest-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=65270. 
Resource load delegate should be able to block poster loading.


Testing before setting poster
EXPECTED (video.clientWidth == '300') OK
EXPECTED (video.clientHeight == '150') OK

Test after setting poster
EXPECTED (video.clientWidth == '300'), OBSERVED '76' FAIL
EXPECTED (video.clientHeight == '150'), OBSERVED '103' FAIL
END OF TEST
